refactor: 컴포넌트, 컨텍스트, 모델 분리
