+++ packages/assignment/src/models/index.ts
@@ -1,0 +1,3 @@
+export * from "./Item";
+export * from "./User";
+export * from "./Notification";
